requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
supabase==2.7.4
pinecone==5.4.0
pandas==2.2.2
numpy==2.1.1
scikit-learn==1.5.2
joblib==1.4.2